Add test for issue 451
